Move repositories into designated directory
