+++ app/components/ui/post-details.tsx
@@ -8,7 +8,7 @@ import {
 	type CommentTreeState,
 	type PostWithOSize,
 } from '#app/types/api-types.ts'
-import { defaultVoteState } from '#app/vote.ts'
+import { defaultVoteState } from '#app/repositories/vote.ts'
 import { PostActionBar } from './post-action-bar.tsx'
 import { PostContent } from './post-content.tsx'
 import { PostInfoBar } from './post-info-bar.tsx'
