+++ app/repositories/ranking.ts
@@ -8,8 +8,8 @@ import {
 	type FrontPagePost,
 } from '#app/types/api-types.ts'
 import { MAX_POSTS_PER_PAGE } from '#app/constants.ts'
-import { type DB } from './types/kysely-types.ts'
-import { type DBEffect } from './types/db-types.ts'
+import { type DB } from '../types/kysely-types.ts'
+import { type DBEffect } from '../types/db-types.ts'
 import {
 	getDescendantCount,
 	getDescendants,
@@ -17,7 +17,7 @@ import {
 	getReplyIds,
 	getPostWithOSizeAndScore,
 } from './post.ts'
-import { relativeEntropy } from './utils/entropy.ts'
+import { relativeEntropy } from '../utils/entropy.ts'
 import { defaultVoteState, getUserVotes } from './vote.ts'
 
 export function toImmutableReplyTree(replyTree: ReplyTree): ImmutableReplyTree {
