+++ app/routes/createPost.tsx
@@ -3,7 +3,7 @@ import { redirect } from '@remix-run/server-runtime'
 import { z } from 'zod'
 import { zfd } from 'zod-form-data'
 import { db } from '#app/db.ts'
-import { createPost } from '#app/post.ts'
+import { createPost } from '#app/repositories/post.ts'
 import { requireUserId } from '#app/utils/auth.server.ts'
 
 const postDataSchema = zfd.formData({
