+++ app/routes/stats.$postId.tsx
@@ -6,8 +6,8 @@ import { type StatsPost } from '#app/types/api-types.ts'
 import { GeneralErrorBoundary } from '#app/components/error-boundary.tsx'
 import { Markdown } from '#app/components/markdown.tsx'
 import { db } from '#app/db.ts'
-import { getStatsPost } from '#app/post.ts'
-import { getEffects } from '#app/ranking.ts'
+import { getStatsPost } from '#app/repositories/post.ts'
+import { getEffects } from '#app/repositories/ranking.ts'
 import { relativeEntropy } from '#app/utils/entropy.ts'
 
 const postIdSchema = z.coerce.number()
