+++ other/replay-vote-events.ts
@@ -1,6 +1,6 @@
 import { type DBVoteEvent } from '../app/types/db-types.ts'
 import { db } from '../app/db.ts'
-import { sendVoteEvent } from '../app/globalbrain.ts'
+import { sendVoteEvent } from '../app/repositories/globalbrain.ts'
 
 async function replayVoteEvents() {
 	console.log('Replaying vote events')
